+++ pkg/runtime/runtime.go
@@ -27,6 +27,7 @@ import (
 	"github.com/docker/docker-agent/pkg/hooks"
 	"github.com/docker/docker-agent/pkg/model/provider"
 	"github.com/docker/docker-agent/pkg/model/provider/options"
+	"github.com/docker/docker-agent/pkg/modelerrors"
 	"github.com/docker/docker-agent/pkg/modelsdev"
 	"github.com/docker/docker-agent/pkg/permissions"
 	"github.com/docker/docker-agent/pkg/rag"
@@ -1173,7 +1174,7 @@ func (r *LocalRuntime) RunStream(ctx context.Context, sess *session.Session) <-c
 				// Auto-recovery: if the error is a context overflow and
 				// session compaction is enabled, compact the conversation
 				// and retry the request instead of surfacing raw errors.
-				var ctxOverflow *ContextOverflowError
+				var ctxOverflow *modelerrors.ContextOverflowError
 				if errors.As(err, &ctxOverflow) && r.sessionCompaction {
 					slog.Warn("Context window overflow detected, attempting auto-compaction",
 						"agent", a.Name(),
@@ -1200,7 +1201,7 @@ func (r *LocalRuntime) RunStream(ctx context.Context, sess *session.Session) <-c
 				slog.Error("All models failed", "agent", a.Name(), "error", err)
 				// Track error in telemetry
 				telemetry.RecordError(ctx, err.Error())
-				events <- Error(formatModelError(err))
+				events <- Error(modelerrors.FormatError(err))
 				streamSpan.End()
 				return
 			}
@@ -2037,312 +2038,6 @@ func (r *LocalRuntime) startSpan(ctx context.Context, name string, opts ...trace
 	return r.tracer.Start(ctx, name, opts...)
 }
 
-// CurrentAgentSubAgentNames implements agenttool.Runner.
-func (r *LocalRuntime) CurrentAgentSubAgentNames() []string {
-	a := r.CurrentAgent()
-	if a == nil {
-		return nil
-	}
-	var names []string
-	for _, sa := range a.SubAgents() {
-		names = append(names, sa.Name())
-	}
-	return names
-}
-
-// RunAgent implements agenttool.Runner. It starts a sub-agent synchronously and
-// blocks until completion or cancellation.
-func (r *LocalRuntime) RunAgent(ctx context.Context, params agenttool.RunParams) *agenttool.RunResult {
-	child, err := r.team.Agent(params.AgentName)
-	if err != nil {
-		return &agenttool.RunResult{ErrMsg: fmt.Sprintf("agent %q not found: %s", params.AgentName, err)}
-	}
-
-	systemMsg := "You are a member of a team of agents. Your goal is to complete the following task:"
-	systemMsg += fmt.Sprintf("\n\n<task>\n%s\n</task>", params.Task)
-	if params.ExpectedOutput != "" {
-		systemMsg += fmt.Sprintf("\n\n<expected_output>\n%s\n</expected_output>", params.ExpectedOutput)
-	}
-
-	sess := params.ParentSession
-
-	// Background tasks run with tools pre-approved because there is no user present
-	// to respond to interactive approval prompts during async execution. This is a
-	// deliberate design trade-off: the user implicitly authorises all tool calls made
-	// by the sub-agent when they approve run_background_agent. Callers should be aware
-	// that prompt injection in the sub-agent's context could exploit this gate-bypass.
-	//
-	// TODO: propagate the parent session's per-tool permission rules once the runtime
-	// supports per-session permission scoping rather than a single shared ToolsApproved flag.
-	s := session.New(
-		session.WithSystemMessage(systemMsg),
-		session.WithImplicitUserMessage("Please proceed."),
-		session.WithMaxIterations(child.MaxIterations()),
-		session.WithTitle("Background agent task"),
-		session.WithToolsApproved(true),
-		session.WithThinking(sess.Thinking),
-		session.WithSendUserMessage(false),
-		session.WithParentID(sess.ID),
-		session.WithAgentName(params.AgentName),
-	)
-
-	var errMsg string
-	events := r.RunStream(ctx, s)
-	for event := range events {
-		if ctx.Err() != nil {
-			break
-		}
-		if choice, ok := event.(*AgentChoiceEvent); ok && choice.Content != "" {
-			if params.OnContent != nil {
-				params.OnContent(choice.Content)
-			}
-		}
-		if errEvt, ok := event.(*ErrorEvent); ok {
-			errMsg = errEvt.Error
-			break
-		}
-	}
-	// Drain remaining events so the RunStream goroutine can complete
-	// and close the channel without blocking on a full buffer.
-	for range events {
-	}
-
-	if errMsg != "" {
-		return &agenttool.RunResult{ErrMsg: errMsg}
-	}
-
-	result := s.GetLastAssistantMessageContent()
-	sess.AddSubSession(s)
-	return &agenttool.RunResult{Result: result}
-}
-
-func (r *LocalRuntime) handleTaskTransfer(ctx context.Context, sess *session.Session, toolCall tools.ToolCall, evts chan Event) (*tools.ToolCallResult, error) {
-	var params struct {
-		Agent          string `json:"agent"`
-		Task           string `json:"task"`
-		ExpectedOutput string `json:"expected_output"`
-	}
-
-	if err := json.Unmarshal([]byte(toolCall.Function.Arguments), &params); err != nil {
-		return nil, fmt.Errorf("invalid arguments: %w", err)
-	}
-
-	a := r.CurrentAgent()
-
-	// Validate that the target agent is in the current agent's sub-agents list
-	subAgents := a.SubAgents()
-	if !slices.ContainsFunc(subAgents, func(sa *agent.Agent) bool { return sa.Name() == params.Agent }) {
-		var subAgentNames []string
-		for _, sa := range subAgents {
-			subAgentNames = append(subAgentNames, sa.Name())
-		}
-		var errorMsg string
-		if len(subAgentNames) > 0 {
-			errorMsg = fmt.Sprintf("Agent %s cannot transfer task to %s: target agent not in sub-agents list. Available sub-agent IDs are: %s", a.Name(), params.Agent, strings.Join(subAgentNames, ", "))
-		} else {
-			errorMsg = fmt.Sprintf("Agent %s cannot transfer task to %s: target agent not in sub-agents list. This agent has no sub-agents configured.", a.Name(), params.Agent)
-		}
-		return tools.ResultError(errorMsg), nil
-	}
-
-	// Span for task transfer (optional)
-	ctx, span := r.startSpan(ctx, "runtime.task_transfer", trace.WithAttributes(
-		attribute.String("from.agent", a.Name()),
-		attribute.String("to.agent", params.Agent),
-		attribute.String("session.id", sess.ID),
-	))
-	defer span.End()
-
-	slog.Debug("Transferring task to agent", "from_agent", a.Name(), "to_agent", params.Agent, "task", params.Task)
-
-	ca := r.CurrentAgentName()
-
-	// Emit agent switching start event
-	evts <- AgentSwitching(true, ca, params.Agent)
-
-	r.setCurrentAgent(params.Agent)
-	defer func() {
-		r.setCurrentAgent(ca)
-
-		// Emit agent switching end event
-		evts <- AgentSwitching(false, params.Agent, ca)
-
-		// Restore original agent info in sidebar
-		if originalAgent, err := r.team.Agent(ca); err == nil {
-			evts <- AgentInfo(originalAgent.Name(), getAgentModelID(originalAgent), originalAgent.Description(), originalAgent.WelcomeMessage())
-		}
-	}()
-
-	// Emit agent info for the new agent
-	if newAgent, err := r.team.Agent(params.Agent); err == nil {
-		evts <- AgentInfo(newAgent.Name(), getAgentModelID(newAgent), newAgent.Description(), newAgent.WelcomeMessage())
-	}
-
-	memberAgentTask := "You are a member of a team of agents. Your goal is to complete the following task:"
-	memberAgentTask += fmt.Sprintf("\n\n<task>\n%s\n</task>", params.Task)
-	if params.ExpectedOutput != "" {
-		memberAgentTask += fmt.Sprintf("\n\n<expected_output>\n%s\n</expected_output>", params.ExpectedOutput)
-	}
-
-	slog.Debug("Creating new session with parent session", "parent_session_id", sess.ID, "tools_approved", sess.ToolsApproved, "thinking", sess.Thinking)
-
-	child, err := r.team.Agent(params.Agent)
-	if err != nil {
-		return nil, err
-	}
-
-	s := session.New(
-		session.WithSystemMessage(memberAgentTask),
-		session.WithImplicitUserMessage("Please proceed."),
-		session.WithMaxIterations(child.MaxIterations()),
-		session.WithTitle("Transferred task"),
-		session.WithToolsApproved(sess.ToolsApproved),
-		session.WithThinking(sess.Thinking),
-		session.WithSendUserMessage(false),
-		session.WithParentID(sess.ID),
-	)
-
-	return r.runSubSession(ctx, sess, s, span, evts, a.Name())
-}
-
-// runSubSession runs a child session within the parent, forwarding events and
-// propagating state (tool approvals, thinking) back to the parent when done.
-func (r *LocalRuntime) runSubSession(ctx context.Context, parent, child *session.Session, span trace.Span, evts chan Event, agentName string) (*tools.ToolCallResult, error) {
-	for event := range r.RunStream(ctx, child) {
-		evts <- event
-		if errEvent, ok := event.(*ErrorEvent); ok {
-			span.RecordError(fmt.Errorf("%s", errEvent.Error))
-			span.SetStatus(codes.Error, "sub-session error")
-			return nil, fmt.Errorf("%s", errEvent.Error)
-		}
-	}
-
-	parent.ToolsApproved = child.ToolsApproved
-	parent.Thinking = child.Thinking
-
-	parent.AddSubSession(child)
-	evts <- SubSessionCompleted(parent.ID, child, agentName)
-
-	span.SetStatus(codes.Ok, "sub-session completed")
-	return tools.ResultSuccess(child.GetLastAssistantMessageContent()), nil
-}
-
-func (r *LocalRuntime) handleHandoff(_ context.Context, _ *session.Session, toolCall tools.ToolCall, _ chan Event) (*tools.ToolCallResult, error) {
-	var params builtin.HandoffArgs
-	if err := json.Unmarshal([]byte(toolCall.Function.Arguments), &params); err != nil {
-		return nil, fmt.Errorf("invalid arguments: %w", err)
-	}
-
-	ca := r.CurrentAgentName()
-	currentAgent, err := r.team.Agent(ca)
-	if err != nil {
-		return nil, fmt.Errorf("current agent not found: %w", err)
-	}
-
-	// Validate that the target agent is in the current agent's handoffs list
-	handoffs := currentAgent.Handoffs()
-	if !slices.ContainsFunc(handoffs, func(a *agent.Agent) bool { return a.Name() == params.Agent }) {
-		var handoffNames []string
-		for _, h := range handoffs {
-			handoffNames = append(handoffNames, h.Name())
-		}
-		var errorMsg string
-		if len(handoffNames) > 0 {
-			errorMsg = fmt.Sprintf("Agent %s cannot hand off to %s: target agent not in handoffs list. Available handoff agent IDs are: %s", ca, params.Agent, strings.Join(handoffNames, ", "))
-		} else {
-			errorMsg = fmt.Sprintf("Agent %s cannot hand off to %s: target agent not in handoffs list. This agent has no handoff agents configured.", ca, params.Agent)
-		}
-		return tools.ResultError(errorMsg), nil
-	}
-
-	next, err := r.team.Agent(params.Agent)
-	if err != nil {
-		return nil, err
-	}
-
-	r.setCurrentAgent(next.Name())
-	handoffMessage := "The agent " + ca + " handed off the conversation to you. " +
-		"Your available handoff agents and tools are specified in the system messages that follow. " +
-		"Only use those capabilities - do not attempt to use tools or hand off to agents that you see " +
-		"in the conversation history from previous agents, as those were available to different agents " +
-		"with different capabilities. Look at the conversation history for context, but only use the " +
-		"handoff agents and tools that are listed in your system messages below. " +
-		"Complete your part of the task and hand off to the next appropriate agent in your workflow " +
-		"(if any are available to you), or respond directly to the user if you are the final agent."
-	return tools.ResultSuccess(handoffMessage), nil
-}
-
-// findModelPickerTool returns the ModelPickerTool from the current agent's
-// toolsets, or nil if the agent has no model_picker configured.
-func (r *LocalRuntime) findModelPickerTool() *builtin.ModelPickerTool {
-	currentName := r.CurrentAgentName()
-	a, err := r.team.Agent(currentName)
-	if err != nil {
-		return nil
-	}
-	for _, ts := range a.ToolSets() {
-		if mpt, ok := tools.As[*builtin.ModelPickerTool](ts); ok {
-			return mpt
-		}
-	}
-	return nil
-}
-
-// handleChangeModel handles the change_model tool call by switching the current agent's model.
-func (r *LocalRuntime) handleChangeModel(ctx context.Context, _ *session.Session, toolCall tools.ToolCall, events chan Event) (*tools.ToolCallResult, error) {
-	var params builtin.ChangeModelArgs
-	if err := json.Unmarshal([]byte(toolCall.Function.Arguments), &params); err != nil {
-		return nil, fmt.Errorf("invalid arguments: %w", err)
-	}
-
-	if params.Model == "" {
-		return tools.ResultError("model parameter is required"), nil
-	}
-
-	// Validate the requested model against the allowed list
-	mpt := r.findModelPickerTool()
-	if mpt == nil {
-		return tools.ResultError("model_picker is not configured for this agent"), nil
-	}
-	allowed := mpt.AllowedModels()
-	if !slices.Contains(allowed, params.Model) {
-		return tools.ResultError(fmt.Sprintf(
-			"model %q is not in the allowed list. Available models: %s",
-			params.Model, strings.Join(allowed, ", "),
-		)), nil
-	}
-
-	return r.setModelAndEmitInfo(ctx, params.Model, events)
-}
-
-// handleRevertModel handles the revert_model tool call by reverting the current agent to its default model.
-func (r *LocalRuntime) handleRevertModel(ctx context.Context, _ *session.Session, _ tools.ToolCall, events chan Event) (*tools.ToolCallResult, error) {
-	return r.setModelAndEmitInfo(ctx, "", events)
-}
-
-// setModelAndEmitInfo sets the model for the current agent and emits an updated
-// AgentInfo event so the UI reflects the change. An empty modelRef reverts to
-// the agent's default model.
-func (r *LocalRuntime) setModelAndEmitInfo(ctx context.Context, modelRef string, events chan Event) (*tools.ToolCallResult, error) {
-	currentName := r.CurrentAgentName()
-	if err := r.SetAgentModel(ctx, currentName, modelRef); err != nil {
-		return tools.ResultError(fmt.Sprintf("failed to set model: %v", err)), nil
-	}
-
-	if a, err := r.team.Agent(currentName); err == nil {
-		events <- AgentInfo(a.Name(), r.getEffectiveModelID(a), a.Description(), a.WelcomeMessage())
-	} else {
-		slog.Warn("Failed to retrieve agent after model change; UI may not reflect the update", "agent", currentName, "error", err)
-	}
-
-	if modelRef == "" {
-		slog.Info("Model reverted via model_picker tool", "agent", currentName)
-		return tools.ResultSuccess("Model reverted to the agent's default model"), nil
-	}
-	slog.Info("Model changed via model_picker tool", "agent", currentName, "model", modelRef)
-	return tools.ResultSuccess(fmt.Sprintf("Model changed to %s", modelRef)), nil
-}
-
 // Summarize generates a summary for the session based on the conversation history.
 // The additionalPrompt parameter allows users to provide additional instructions
 // for the summarization (e.g., "focus on code changes" or "include action items").
@@ -2487,24 +2182,4 @@ func estimateMessageTokens(msg *chat.Message) int64 {
 	}
 
 	return int64(chars/charsPerToken) + perMessageOverhead
-}
-
-// formatModelError produces a user-friendly error message from a model error.
-// Raw HTTP errors with request IDs, JSON payloads, and API URLs are replaced
-// with actionable guidance. Context overflow errors receive a dedicated
-// message; other errors are cleaned up to remove noise while preserving the
-// essential failure reason.
-func formatModelError(err error) string {
-	if err == nil {
-		return ""
-	}
-
-	// Context overflow gets a dedicated, actionable message.
-	var ctxOverflow *ContextOverflowError
-	if errors.As(err, &ctxOverflow) {
-		return "The conversation has exceeded the model's context window and automatic compaction is not enabled. " +
-			"Try running /compact to reduce the conversation size, or start a new session."
-	}
-
-	return err.Error()
 }
